requirements.txt
django==5.1.1
Pillow==10.4.0
ruff
ipdb
black
djangorestframework
